fix: resolve compiler warnings for size_t to uint32_t conversion
- Changed m_dead_count and m_wasted_string_bytes to size_t to match 64-bit platform requirements.
- Cleaned up potential precision loss in string pool and dead entry accounting.

+++ src/mft/MftReader.h
@@ -111,8 +111,8 @@ private:
     mutable QReadWriteLock m_dataLock;
     bool m_isInitialized = false;
     uint32_t m_dirty_count = 0;
-    uint32_t m_dead_count = 0;
-    uint32_t m_wasted_string_bytes = 0;
+    size_t   m_dead_count = 0;
+    size_t   m_wasted_string_bytes = 0;
     std::vector<uint32_t> m_sorted_indices;
 };
